+++ src/main/java/com/grash/dto/LaborPatchDTO.java
@@ -14,7 +14,7 @@ public class LaborPatchDTO {
 
     private boolean includeToTotalTime;
 
-    private double hourlyRate;
+    private long hourlyRate;
 
     private int duration;
