+++ src/main/java/com/grash/model/Location.java
@@ -31,9 +31,9 @@ public class Location extends CompanyAudit {
 
     private String address;
 
-    private Double longitude;
+    private Long longitude;
 
-    private Double latitude;
+    private Long latitude;
 
     private boolean hasChildren;
